+++ include/aikido/constraint/TestableIntersection.hpp
@@ -23,7 +23,12 @@ public:
 
   // Documentation inherited.
   bool isSatisfied(
-      const aikido::statespace::StateSpace::State* state) const override;
+      const aikido::statespace::StateSpace::State* state,
+      TestableOutcome* outcome = nullptr) const override;
+
+  /// Return an instance of DefaultTestableOutcome, since this class doesn't
+  /// have a more specialized TestableOutcome derivative assigned to it.
+  std::unique_ptr<TestableOutcome> createOutcome() const override;
 
   // Documentation inherited.
   statespace::StateSpacePtr getStateSpace() const override;