+++ include/aikido/statespace/dart/MetaSkeletonStateSpace.hpp
@@ -164,6 +164,8 @@ private:
 };
 
 using MetaSkeletonStateSpacePtr = std::shared_ptr<MetaSkeletonStateSpace>;
+using ConstMetaSkeletonStateSpacePtr
+    = std::shared_ptr<const MetaSkeletonStateSpace>;
 
 } // namespace dart
 } // namespace statespace